(maint) add deadcode check
