+++ tracing/main.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	_ "embed"
 	"fmt"
-	"os"
-	"runtime/debug"
 	"time"
 
 	"github.com/getsentry/sentry-go"
@@ -190,51 +188,4 @@ func (h *UserAgentSampler) ShouldSample(parameters sdktrace.SamplingParameters) 
 // Description returns information describing the Sampler.
 func (h *UserAgentSampler) Description() string {
 	return "Simple Sampler based on the UserAgent of the request"
-}
-
-// LogRecoverToReturn Recovers from a panic, logs and forwards it sentry and otel, then returns
-// Does nothing when there is no panic.
-func LogRecoverToReturn(ctx context.Context, loc string) {
-	err := recover()
-	if err == nil {
-		return
-	}
-
-	stack := string(debug.Stack())
-	handleError(ctx, loc, err, stack)
-}
-
-// LogRecoverToExit Recovers from a panic, logs and forwards it sentry and otel, then exits
-// Does nothing when there is no panic.
-func LogRecoverToExit(ctx context.Context, loc string) {
-	err := recover()
-	if err == nil {
-		return
-	}
-
-	stack := string(debug.Stack())
-	handleError(ctx, loc, err, stack)
-
-	// ensure that errors still get sent out
-	ShutdownTracer()
-
-	os.Exit(1)
-}
-
-func handleError(ctx context.Context, loc string, err interface{}, stack string) {
-	msg := fmt.Sprintf("unhandled panic in %v, exiting: %v", loc, err)
-
-	hub := sentry.CurrentHub()
-	if hub != nil {
-		hub.Recover(err)
-	}
-
-	if ctx != nil {
-		log.WithContext(ctx).WithFields(log.Fields{"loc": loc, "stack": stack}).Error(msg)
-		span := trace.SpanFromContext(ctx)
-		span.SetAttributes(attribute.String("ovm.panic.loc", loc))
-		span.SetAttributes(attribute.String("ovm.panic.stack", stack))
-	} else {
-		log.WithFields(log.Fields{"loc": loc, "stack": stack}).Error(msg)
-	}
 }
